Refactor(docs): Extract Placement utils
