+++ lib/docs/src/utils/placement.ts
@@ -1,0 +1,19 @@
+import { PopoverPlacement } from "../../../core/src";
+
+export const PLACEMENTS: PopoverPlacement[] = [
+	"top",
+	"top-start",
+	"top-end",
+	"right",
+	"right-start",
+	"right-end",
+	"bottom",
+	"bottom-start",
+	"bottom-end",
+	"left",
+	"left-start",
+	"left-end",
+	"auto",
+	"auto-start",
+	"auto-end",
+];
